feat: construction of new_card with images? (closes #36)
So it's a bit of a hack stemming from card fields wanting to be primitive types (in this case a string filename), on top of that it wants the path to relative to the active set - so you _need_ a set to write image files against. That seems a bit contrary to how new_card wants to operate, but I might be understating its usefulness.